+++ returns/_generated/pipeline/pipe.py
@@ -1,39 +1,7 @@
-from typing import Generic, Tuple, TypeVar
-
-from typing_extensions import final
-
 from returns._generated.pipeline.flow import _flow
-from returns.primitives.types import Immutable
-
-_InstanceType = TypeVar('_InstanceType')
-_PipelineStepType = TypeVar('_PipelineStepType')
-_ReturnType = TypeVar('_ReturnType')
 
 
-@final
-class _Pipe(
-    Generic[_InstanceType, _ReturnType, _PipelineStepType],
-    Immutable,
-):
-    """
-    Internal type to make ``mypy`` plugin work correctly.
-
-    We need this to be able to check ``__call__`` signature.
-    See docs on ``pipe`` feature in ``mypy`` plugin.
-    """
-
-    __slots__ = ('_functions',)
-
-    def __init__(self, functions: Tuple[_PipelineStepType, ...]) -> None:
-        object.__setattr__(self, '_functions', functions)  # noqa: WPS609
-
-    def __call__(self, instance: _InstanceType) -> _ReturnType:
-        return _flow(instance, *self._functions)  # type: ignore
-
-
-def _pipe(
-    *functions: _PipelineStepType,
-) -> _Pipe[_InstanceType, _ReturnType, _PipelineStepType]:
+def _pipe(*functions):
     """
     Allows to compose a value and up to 7 functions that use this value.
 
@@ -63,4 +31,4 @@ def _pipe(
         - https://github.com/gcanti/fp-ts/blob/master/src/pipeable.ts
 
     """
-    return _Pipe(functions)
+    return lambda instance: _flow(instance, *functions)