+++ CHANGELOG.md
@@ -46,6 +46,8 @@ See [0Ver](https://0ver.org/).
 
 - **Breaking**: fixes serious typing issue and changes how `flow` works,
   now it has a hard limit of 21 parameters: 1 instance + 20 functions
+- **Breaking**: fixes serious typing issue and changes how `pipe` works,
+  now it has a hard limit of 20 parameters
 
 ### Misc